Issue #396: treat registry.consul.kvpath as prefix
This patch extends the behavior of the registry.consul.kvpath
and treats it as a prefix instead of a single key. fabio will
now list the key and all available subkeys in alphabetical order
and combine them for the routing table. Each subsection will get
header which contains key name.

Todo: the ui needs to handle that as well